requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-solo==1.1.0
django-suit==0.2.14
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
